clean up Google Analytics related code
