+++ source/js/primary-nav.js
@@ -1,4 +1,4 @@
-import ReactGA from "react-ga";
+import ReactGA from "./react-ga-proxy.js";
 import navNewsletter from "./nav-newsletter.js";
 
 let primaryNav = {
